+++ app/exported.go
@@ -17,6 +17,7 @@ import (
 	server "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/http"
 	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/http/healthcheck"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-sender.git/handlers"
+	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-sender.git/handlers/nonce"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-sender.git/handlers/producer"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-sender.git/handlers/sender"
 )
@@ -64,11 +65,15 @@ func registerHandlers() {
 	// Generic handlers on every worker
 	engine.Register(logger.Logger)
 	engine.Register(loader.Loader)
+
 	engine.Register(offset.Marker)
+	// RecoveryStatusSetter middleware must surround producer (see docstring on RecoveryStatusSetter handler)
+	engine.Register(nonce.GlobalRecoveryStatusSetter())
 	engine.Register(producer.GlobalHandler())
 	engine.Register(opentracing.GlobalHandler())
 
 	// Specific handlers tk Sender worker
+	engine.Register(nonce.GlobalChecker())
 	engine.Register(sender.GlobalHandler())
 }