+++ handlers/producer/producer.go
@@ -12,23 +12,33 @@ import (
 
 // PrepareMsg prepare message to produce from TxContexts
 func PrepareMsg(txctx *engine.TxContext, msg *sarama.ProducerMessage) error {
-	// If an error occurred then we redirect to recovery
+	// Set message Key
+	msg.Key = sarama.StringEncoder(utils.ToChainAccountKey(txctx.Envelope.GetChain().ID(), txctx.Envelope.Sender()))
+
+	// We loop over errors to
+	// - determine to which topic to send transaction
+	// - remove invalid nonce warnings
+loop:
 	for _, err := range txctx.Envelope.GetErrors() {
-		if errors.IsWarning(err) {
+		switch {
+		case errors.IsWarning(err):
 			continue
+		default:
+			// If an error occurred we redirect to recovery
+			msg.Topic = viper.GetString("kafka.topic.recover")
+			break loop
 		}
+	}
 
-		// Marshal Envelope into sarama Message
-		err := encoding.Marshal(txctx.Envelope, msg)
-		if err != nil {
-			return err
-		}
+	// If no error and nonce is invalid we redirect envelope to tx-nonce
+	if b, ok := txctx.Get("invalid.nonce").(bool); len(txctx.Envelope.GetErrors()) == 0 && ok && b {
+		msg.Topic = viper.GetString("kafka.topic.nonce")
+	}
 
-		// Set Topic to Nonce topic
-		msg.Topic = viper.GetString("kafka.topic.recover")
-		msg.Key = sarama.StringEncoder(utils.ToChainAccountKey(txctx.Envelope.GetChain().ID(), txctx.Envelope.Sender()))
-
-		break
+	// Marshal Envelope into sarama Message
+	err := encoding.Marshal(txctx.Envelope, msg)
+	if err != nil {
+		return err
 	}
 
 	return nil