+++ handlers/producer/producer_test.go
@@ -22,4 +22,17 @@ func TestPrepareMsg(t *testing.T) {
 	_ = txctx.Error(errors.ConnectionError("Connection error"))
 	_ = PrepareMsg(txctx, msg)
 	assert.Equal(t, "topic-tx-recover", msg.Topic, "If error out topic should be recovery")
+
+	txctx = engine.NewTxContext()
+	txctx.Envelope = &envelope.Envelope{}
+	txctx.Set("invalid.nonce", true)
+	_ = PrepareMsg(txctx, msg)
+	assert.Equal(t, "topic-tx-nonce", msg.Topic, "If invalid nonce out topic should be tx-nonce")
+
+	txctx = engine.NewTxContext()
+	txctx.Envelope = &envelope.Envelope{}
+	txctx.Set("invalid.nonce", true)
+	_ = txctx.Error(errors.ConnectionError("nonce too low"))
+	_ = PrepareMsg(txctx, msg)
+	assert.Equal(t, "topic-tx-recover", msg.Topic, "If invalid nonce and error topic should be tx-recover")
 }