+++ handlers/store/exported.go
@@ -1,0 +1,3 @@
+package store
+
+const component = "handler.store"