+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Version 0.6.3
 
-*Released on September 26th 2019*
+*Unreleased*
 
+- [REFACTOR] Refactor handler sender into a composite handler
 - [FEAT] Nonce-Management v2
 
 ### Version 0.6.2